pyproject.toml
# pyproject.toml file

[tool.poetry]
name = "threatquery"
version = "0.1.0"
description = ""
authors = ["Emre Eren <emreren@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.6"
httpx = "^0.24.1"
fastapi = {extras = ["all"], version = "^0.100.1"}
sqlalchemy = "^2.0.19"
python-dotenv = "^1.0.0"
logging = "^0.4.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
